Add code to add and remove list through DOM
Fill in the add list form to add a list to a project and remove it by
using the delete button next to each list.
Ps: I am aware of a lot of shortcomings/bugs in my code but they will be
fixed in the future. Most of them will get fixed as the project
progresses and i will fix the others after completing the DOM
functionality.