Fix bug that deleted wrong task when in All ToDos
In previous commit, i changed rendering all to-dos/projects/lists when
deleting one to removing that specific todo/projectV/list from DOM but
since the remaining to-dos/projects/lists are not rendered again, their
indices are not updated and then deleting more to-dos/projects/lists
with outdated indices from the same collection results in incorrect
deletion. Now all the to-dos/projects/lists are rendered again when one
of them is deleted.