Add functionality to edit existing to-do
I also created a separate function for rendering toDos by checking all
conditions. I was thinking of doing this kimd of stuff later but its
probably best to keep your code as optimised/clean as possible. I mean
if i am going to do it later anyways then why not do it on the first
time? It will simply be more efficient that way.