Add init method to DomHandler
The init method initialises the page on load. It makes sure that the
projects (if any) are rendered without having to use the add project
button ans sets first project as currentProject. It also renders all
the lists present in the currentProject and makes 1st list as
currentList and renders its toDos;